ClassNameCheck_ArcgisPro
In the query expression of the attribute table in #ArcGISpro, “fieldname” shows all domain’s (if domain loaded) descriptions (assigned or non-assigned), which is dispensable. To identify the only assigned class in ArcGIS Pro, I developed a script in Python (arcpy). This script will help you select features based on assigned domains, similar to the functionality in ArcGIS Desktop. I think it will be helpful.